backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.8
chainkit==0.1.0
better-auth==0.0.1
qdrant-client==1.8.0
asyncpg==0.29.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-json-logger==2.0.7